Create script
The goal of this script is to simulate the effects of size-specific harvest on the resilience of mangrove species in South Africa based on existing transition population matrices and use age-from stage modeling to investigate the role of aging in the patterns of response to harvest. This script accompanies the paper Gaoue OG & Yessoufou K (2019) Strong seedling recruitment does not limit mangrove vulnerability to harvest. Environmental Research Letters, https://doi.org/10.1088/1748-9326/ab1def